# Improve dictionary performance