--- a/src/asm_writing/rewriter.h
+++ b/src/asm_writing/rewriter.h
@@ -430,6 +430,7 @@ protected:
     int _allocate(RewriterVar* result, int n);
     void _allocateAndCopy(RewriterVar* result, RewriterVar* array, int n);
     void _allocateAndCopyPlus1(RewriterVar* result, RewriterVar* first_elem, RewriterVar* rest, int n_rest);
+    void _checkAndThrowCAPIException(RewriterVar* r);
 
     // The public versions of these are in RewriterVar
     void _addGuard(RewriterVar* var, RewriterVar* val_constant);
@@ -513,6 +514,9 @@ public:
     RewriterVar* allocate(int n);
     RewriterVar* allocateAndCopy(RewriterVar* array, int n);
     RewriterVar* allocateAndCopyPlus1(RewriterVar* first_elem, RewriterVar* rest, int n_rest);
+
+    // This emits `if (!r) throwCAPIException()`
+    void checkAndThrowCAPIException(RewriterVar* r);
 
     void abort();
     void commit();