--- a/src/runtime/objmodel.cpp
+++ b/src/runtime/objmodel.cpp
@@ -3696,7 +3696,7 @@ Box* callCLFunc(CLFunction* f, CallRewriteArgs* rewrite_args, int num_output_arg
         ASSERT(chosen_cf->spec->rtn_type->isFitBy(r->cls), "%s (%p) was supposed to return %s, but gave a %s",
                g.func_addr_registry.getFuncNameAtAddress(chosen_cf->code, true, NULL).c_str(), chosen_cf->code,
                chosen_cf->spec->rtn_type->debugName().c_str(), r->cls->tp_name);
-        assert(!PyErr_Occurred());
+        ASSERT(!PyErr_Occurred(), "%p", chosen_cf->code);
     }
 
     return r;