--- a/src/runtime/objmodel.cpp
+++ b/src/runtime/objmodel.cpp
@@ -2056,12 +2056,11 @@ bool dataDescriptorSetSpecialCases(Box* obj, Box* val, Box* descr, SetattrRewrit
             args.push_back(r_obj);
             args.push_back(r_val);
             args.push_back(r_closure);
-            rewrite_args->rewriter->call(
+            RewriterVar* r_rtn = rewrite_args->rewriter->call(
                 /* has_side_effects */ true, (void*)getset_descr->set, args);
 
             if (descr->cls == capi_getset_cls)
-                // TODO I think we are supposed to check the return value?
-                rewrite_args->rewriter->call(true, (void*)checkAndThrowCAPIException);
+                rewrite_args->rewriter->checkAndThrowCAPIException(r_rtn, -1);
 
             rewrite_args->out_success = true;
         }
@@ -2578,7 +2577,11 @@ template <ExceptionStyle S> BoxedInt* lenInternal(Box* obj, LenRewriteArgs* rewr
             // support calling a RewriterVar (can only call fixed function addresses).
             r_m->addAttrGuard(offsetof(PySequenceMethods, sq_length), (intptr_t)m->sq_length);
             RewriterVar* r_n = rewrite_args->rewriter->call(true, (void*)m->sq_length, r_obj);
-            rewrite_args->rewriter->call(true, (void*)checkAndThrowCAPIException);
+
+            // Some CPython code seems to think that any negative return value means an exception,
+            // but the docs say -1. TODO it would be nice to just handle any negative value.
+            rewrite_args->rewriter->checkAndThrowCAPIException(r_n, -1);
+
             RewriterVar* r_r = rewrite_args->rewriter->call(false, (void*)boxInt, r_n);
 
             rewrite_args->out_success = true;
@@ -4231,7 +4234,8 @@ Box* compareInternal(Box* lhs, Box* rhs, int op_type, CompareRewriteArgs* rewrit
                     // support calling a RewriterVar (can only call fixed function addresses).
                     r_sqm->addAttrGuard(offsetof(PySequenceMethods, sq_contains), (intptr_t)sqm->sq_contains);
                     RewriterVar* r_b = rewrite_args->rewriter->call(true, (void*)sqm->sq_contains, r_rhs, r_lhs);
-                    rewrite_args->rewriter->call(true, (void*)checkAndThrowCAPIException);
+                    rewrite_args->rewriter->checkAndThrowCAPIException(r_b, -1);
+
                     // This could be inlined:
                     RewriterVar* r_r;
                     if (op_type == AST_TYPE::NotIn)