--- a/src/runtime/types.h
+++ b/src/runtime/types.h
@@ -661,7 +661,7 @@ struct PyLt {
 
 class BoxedDict : public Box {
 public:
-    typedef std::unordered_map<Box*, Box*, PyHasher, PyEq, StlCompatAllocator<std::pair<Box*, Box*>>> DictMap;
+    typedef std::unordered_map<Box*, Box*, PyHasher, PyEq> DictMap;
 
     DictMap d;