The backstory is that while we do a pretty decent job of avoiding using dictionaries where CPython does, I was running into some cases where we use them as well, and we're slower since our dictionaries are slower.  This particular example was instantiating namedtuples, since the namedtuple class is created in an exec, and so when we get inside its `__init__`, we have to do slow lookups to access any of the builtins.  We could also try to optimize those cases specifically, or we could make our hidden class logic a "dict storage strategy" or some such.

But anyway, this pr just does some quick&easy things:
- switch from std::unordered_map to llvm::DenseMap
  - The issue here is that llvm::DenseMap doesn't have a version that takes a custom allocator; Chris had #558, but I decided to fold in a change to explicitly scan the dict contents rather than letting the conservative scanner handle it.  This also exposed some other issues, so this ended up being the bulk of this PR by loc.
- Inline the emitted calls to checkAndThrowCAPIException(), which is much more expensive than being able to check the return value directly.

```
      django_template3.py             3.3s (2)             3.2s (2)  -2.2%
            pyxl_bench.py             2.8s (2)             2.8s (2)  -0.5%
sqlalchemy_imperative2.py             3.2s (2)             3.2s (2)  -0.3%
                  geomean                 3.1s                 3.0s  -1.0%
```